Add files via upload
The objective of the project is to make an application in android platform to 
purchase items in an existing shop. In order to build such an application complete web 
support need to be provided. A complete and efficient web application which can provide 
the online shopping experience is the basic objective of the project. The web application can 
be implemented in the form of an android application with web view.